requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dashscope==1.22.1
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
idna==3.10
multidict==6.1.0
pip==25.0.1
propcache==0.2.1
pydub==0.25.1
pypdf==5.3.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
urllib3==2.3.0
websocket-client==1.8.0
wheel==0.45.1
yarl==1.18.3
